Add files via upload
In this Project we are using credit card fraud dataset to detect the fraud or non-fraud transaction using logistic regression. 
Logistic regression is a statistical method that is used for building machine learning models where the dependent variable is dichotomous: i.e. binary. Logistic regression is used to describe data and the relationship between one dependent variable and one or more independent variables.