Add files via upload
This is the titanic survival prediction project where are predicting whether a passenger will be able to survive or not. Gender, age, ticket type, and socioeconomic status all had a role in whether a passenger would be fortunate enough to survive the Titanic or sadly perish, according to a prediction model developed using passenger data. We did the following tasks in the project:
         . Importing dataset
         . Exploratory Data Analysis
         . Data pre-processing
         . Model Fitting
         . Model Evaluation etc.